Beckend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
requests==2.32.3
httpx==0.27.2
pydantic==2.10.6
pydantic-settings==2.7.1
pyTelegramBotAPI==4.30.0
pillow==11.0.0
gunicorn==23.0.0
